Norm check

+++ philo/setters.c
@@ -6,7 +6,7 @@
 /*   By: adbouras <adbouras@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/16 10:10:20 by adbouras          #+#    #+#             */
-/*   Updated: 2024/07/16 11:06:07 by adbouras         ###   ########.fr       */
+/*   Updated: 2024/07/19 10:48:11 by adbouras         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	set_bool(bool *target, pthread_mutex_t *mutex, bool value)
 	mutex_handle(mutex, UNLOCK);
 }
 
-void	set_long(size_t *target, pthread_mutex_t *mutex, size_t value)
+void	set_long(long *target, pthread_mutex_t *mutex, long value)
 {
 	mutex_handle(mutex, LOCK);
 	*target = value;
